requirements.txt
tqdm==4.62.1
seaborn==0.11.1
matplotlib==3.4.2
numpy==1.19.4
scipy==1.5.3
opencv_python==4.4.0.46
tensorflow==2.3.0
scikit_image==0.17.2
Pillow==9.0.0
scikit_learn==1.0.2
skimage==0.0
thispersondoesnotexist==1.0.2
